requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
html5lib==1.1
idna==3.3
pycodestyle==2.8.0
python-environ==0.4.54
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
toml==0.10.2
urllib3==1.26.9
webencodings==0.5.1
